requirements.txt
langchain-openai
langchain-community
langchain-huggingface
langchain-experimental
sentence-transformers
langchain-chroma
faiss-cpu
rank_bm25
fastapi
uvicorn
streamlit
ragas
pymupdf
beautifulsoup4
python-dotenv
tiktoken
pydantic
pydantic-settings
uvicorn
python-multipart
langchain-google-genai
langchain-anthropic


